# Viewer draw better performance