Performance fixes for the viewer draw
One Icosphere with 8 subdivisions before 0.851 s  now 0.251 s
One Icosphere with 8 subdivisions and 5 matrices before 4.517 s  now 3.720 s

- [x] Ready for merge.